Use fromIntegral in Time to support multiple versions of clock

+++ src/Wrench/Time.hs
@@ -33,4 +33,4 @@ toSeconds (TimeDelta t) = t
 getTicks :: IO TimeTicks
 getTicks = do
   (TimeSpec s ns) <- getTime Monotonic
-  return $ tickSeconds s + tickNanoSeconds ns
+  return $ tickSeconds s + tickNanoSeconds (fromIntegral ns)
